+++ ReactNativeTraining/src/app/screens/components/Chat/styles.js
