+++ ReactNativeTraining/src/app/screens/components/ContactList/index.js
@@ -7,12 +7,15 @@ import styles from './styles';
 
 export default class ContactList extends Component {
   separatorComponent = () => <View style={styles.greyLine} />;
+
+  handlePress = item => this.props.onClick(item);
+
   renderItem = item => (
     <ListItem
       hideChevron
       roundAvatar
       avatar={item.avatar}
-      onPress={() => this.props.onClick(item)}
+      onPress={() => this.handlePress(item)}
       title={item.name}
       titleStyle={styles.name}
       style={styles.listItem}
