+++ ReactNativeTraining/src/app/screens/components/Message/index.js
