+++ ReactNativeTraining/src/redux/chat/actions.js
@@ -1,0 +1,3 @@
+export const chatSelected = id => dispatch => dispatch({ type: 'CHAT_SELECTED', payload: id });
+
+export default chatSelected;
