+++ ReactNativeTraining/src/redux/chat/reducer.js
@@ -1,0 +1,8 @@
+export const chat = (state = [], action) => {
+  switch (action.type) {
+    case 'CHAT_SELECTED':
+      return action.payload;
+    default:
+      return state;
+  }
+};
