+++ ReactNativeTraining/src/redux/contacts/reducer.js
@@ -1,0 +1,8 @@
+export const contacts = (state = [], action) => {
+  switch (action.type) {
+    case 'LOAD_CONTACTS':
+      return action.payload;
+    default:
+      return state;
+  }
+};
