+++ ReactNativeTraining/src/redux/groups/reducer.js
@@ -1,0 +1,8 @@
+export const groups = (state = [], action) => {
+  switch (action.type) {
+    case 'LOAD_GROUPS':
+      return action.payload;
+    default:
+      return state;
+  }
+};
