+++ ReactNativeTraining/src/redux/messages/reducer.js
@@ -1,0 +1,8 @@
+export const messages = (state = [], action) => {
+  switch (action.type) {
+    case 'LOAD_MESSAGES':
+      return action.payload;
+    default:
+      return state;
+  }
+};
